feat: separates widgets and layouts into modules
