+++ Cargo.toml
@@ -18,7 +18,8 @@ default = ["web"]
 
 [workspace]
 members = [
-  "web",
+  "crates/web",
+  "crates/macro",
   "examples/hello-world",
 ]
