+++ crates/macro/Cargo.toml
@@ -1,14 +1,9 @@
 [package]
-name = "hello-world"
+name = "elvis-macro"
 version = "0.1.0"
 authors = ["clearloop <udtrokia@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
-[lib]
-crate-type = ["cdylib"]
-
 [dependencies]
-wasm-bindgen = "0.2.60"
-elvis-web = { path = "../../crates/web" }
