+++ crates/web/Cargo.toml
@@ -17,7 +17,7 @@ wasm-bindgen = "0.2.60"
 js-sys = "0.3.37"
 
 [dependencies.elvis]
-path = "../"
+path = "../.."
 version = "0.2.1"
 features = ["web"]
