+++ crates/web/src/enums.rs
