+++ src/bridge/macros.rs
@@ -1,59 +1,4 @@
-//! Convert widgets to Tree
-use crate::{
-    Align, Center, Col, Container, Flex, Grid, Image, List, MultiColumn, Row, Serde, SizedBox,
-    Text, Tree,
-};
-use std::{cell::RefCell, collections::HashMap, convert::Into, rc::Rc};
-
-// widgets
-impl<'t> Into<Tree> for &'t Text {
-    fn into(self) -> Tree {
-        let mut m = HashMap::<String, String>::new();
-        let mut cm = HashMap::<String, String>::new();
-
-        m.insert("style".into(), self.style.ser());
-        cm.insert("text".into(), self.text.to_string());
-
-        Tree::new(
-            m,
-            vec![Tree::new(cm, vec![], None, "plain".into())],
-            None,
-            "p".into(),
-        )
-        .borrow()
-        .to_owned()
-    }
-}
-
-impl<'i> Into<Tree> for &'i Image {
-    fn into(self) -> Tree {
-        let mut m = HashMap::<String, String>::new();
-        m.insert("class".into(), "elvis-image".into());
-        m.insert("style".into(), self.src.ser());
-
-        Tree::new(
-            m,
-            vec![Rc::new(RefCell::new(self.child.to_owned()))],
-            None,
-            "div".into(),
-        )
-        .borrow()
-        .to_owned()
-    }
-}
-
-// layouts
-impl<'i> Into<Tree> for &'i Center {
-    fn into(self) -> Tree {
-        let mut m = HashMap::<String, String>::new();
-        m.insert("class".into(), "elvis-center elvis-flex".into());
-
-        let mut cs = vec![];
-
-        cs.push(Rc::new(RefCell::new(self.child.to_owned())));
-        Tree::new(m, cs, None, "div".into()).borrow().to_owned()
-    }
-}
+#![macro_use]
 
 /// multi-child widget
 macro_rules! mcw {
@@ -150,31 +95,4 @@ macro_rules! sw {
             }
         )*
     };
-}
-
-sw! {
-    Align,
-    Container,
-    Flex,
-    SizedBox,
-}
-
-mcw! {
-    Col,
-    List,
-    Row,
-}
-
-mcws! {
-    Grid,
-    MultiColumn,
-}
-
-it! {
-    Center,
-    Col,
-    Row,
-    Image,
-    Text,
-    List,
 }
