+++ src/values/mod.rs
@@ -1,3 +1,5 @@
-pub mod color;
+mod color;
 pub mod layout;
-pub mod unit;
+mod unit;
+
+pub use {color::Colors, unit::Unit};
