+++ src/widgets/mod.rs
@@ -1,0 +1,7 @@
+mod image;
+mod text;
+
+pub use {
+    image::{Image, ImageSrc},
+    text::{Text, TextStyle},
+};
