+++ src/widgets/text.rs
@@ -1,4 +1,4 @@
-use crate::{Colors, Tree, Unit};
+use crate::values::{Colors, Unit};
 
 /// `Text` might be the most popular spider from Mars,
 /// Does it know the Great Ziggy Stardust?
@@ -59,30 +59,5 @@ impl Default for TextStyle {
             height: Unit::Rem(1.0),
             stretch: Unit::Percent(100.0),
         }
-    }
-}
-
-/// If you don't want Image playing in background anonymously, just remove its child.
-pub struct Image {
-    pub src: ImageSrc,
-    pub child: Tree,
-}
-
-impl Image {
-    /// image will auto-fill to the father widget
-    pub fn new(src: String, child: Tree) -> Image {
-        Image {
-            src: ImageSrc(src),
-            child,
-        }
-    }
-}
-
-/// Image src
-pub struct ImageSrc(pub String);
-
-impl ImageSrc {
-    pub fn as_bytes(&self) -> &[u8] {
-        self.0.as_bytes()
     }
 }
